+++ src/providers/seventv/SeventvEmotes.hpp
@@ -40,6 +40,7 @@ public:
 
     std::shared_ptr<const EmoteMap> emotes() const;
     boost::optional<EmotePtr> emote(const EmoteName &name) const;
+    static void addEmote(QString emoteID, TwitchChannel *channel);
     void loadEmotes();
     static void loadChannel(std::weak_ptr<Channel> channel,
                             const QString &channelId,