+++ src/singletons/Settings.hpp
@@ -185,11 +185,11 @@ public:
     BoolSetting enableHomiesCompletion = {
         "/behaviour/autocompletion/enableHomiesCompletion", true};
     BoolSetting enableFFZCompletion = {
-            "/behaviour/autocompletion/enableFFZCompletion", true};
+        "/behaviour/autocompletion/enableFFZCompletion", true};
     BoolSetting enableBTTVCompletion = {
-            "/behaviour/autocompletion/enableBTTVCompletion", true};
+        "/behaviour/autocompletion/enableBTTVCompletion", true};
     BoolSetting enable7TVCompletion = {
-            "/behaviour/autocompletion/enable7TVCompletion", true};
+        "/behaviour/autocompletion/enable7TVCompletion", true};
 
     FloatSetting pauseOnHoverDuration = {"/behaviour/pauseOnHoverDuration", 0};
     EnumSetting<Qt::KeyboardModifier> pauseChatModifier = {
@@ -218,12 +218,10 @@ public:
 
     BoolSetting enableHomiesGlobalEmotes = {"/emotes/enableHomiesGlobalEmotes",
                                             true};
-    BoolSetting enable7TVGlobalEmotes = {"/emotes/enable7TVGlobalEmotes",
-                                            true};
-    BoolSetting enableFFZGlobalEmotes = {"/emotes/enableFFZGlobalEmotes",
-                                            true};
+    BoolSetting enable7TVGlobalEmotes = {"/emotes/enable7TVGlobalEmotes", true};
+    BoolSetting enableFFZGlobalEmotes = {"/emotes/enableFFZGlobalEmotes", true};
     BoolSetting enableBTTVGlobalEmotes = {"/emotes/enableBTTVGlobalEmotes",
-                                            true};
+                                          true};
 
     /// Links
     BoolSetting linksDoubleClickOnly = {"/links/doubleClickToOpen", false};