+++ src/widgets/settingspages/HomiesPage.cpp
@@ -7,9 +7,10 @@
 #include "Application.hpp"
 #include "common/Version.hpp"
 #include "widgets/BaseWindow.hpp"
+#include "widgets/dialogs/SelectChannelSeparateLinksDialog.hpp"
+#include "widgets/dialogs/TokenDialog.hpp"
 #include "widgets/helper/Line.hpp"
 #include "widgets/settingspages/GeneralPageView.hpp"
-#include "widgets/dialogs/SelectChannelSeparateLinksDialog.hpp"
 
 #include <QDesktopServices>
 #include <QFileDialog>
@@ -43,6 +44,51 @@ void HomiesPage::initLayout(GeneralPageView &layout)
 {
     auto &s = *getSettings();
 
+    QPushButton *sevenTVTokenButton = new QPushButton();
+    {
+        sevenTVTokenButton->setText("7TV Authentication Token");
+        sevenTVTokenButton->adjustSize();
+    }
+
+    connect(sevenTVTokenButton, &QPushButton::clicked, [=]() {
+        auto SevenTVTokenWidget = new TokenDialog(0);
+
+        SevenTVTokenWidget->show();
+        SevenTVTokenWidget->raise();
+    });
+
+    QPushButton *FFZTokenButton = new QPushButton();
+    {
+        FFZTokenButton->setText("FFZ Authentication Token");
+        FFZTokenButton->adjustSize();
+    }
+
+    connect(FFZTokenButton, &QPushButton::clicked, [=]() {
+        auto FFZTokenWidget = new TokenDialog(1);
+
+        FFZTokenWidget->show();
+        FFZTokenWidget->raise();
+    });
+
+    QPushButton *BTTVTokenButton = new QPushButton();
+    {
+        BTTVTokenButton->setText("BetterTTV Authentication Token");
+        BTTVTokenButton->adjustSize();
+    }
+
+    connect(BTTVTokenButton, &QPushButton::clicked, [=]() {
+        auto BTTVTokenWidget = new TokenDialog(2);
+
+        BTTVTokenWidget->show();
+        BTTVTokenWidget->raise();
+    });
+
+    layout.addTitle("Providers tokens");
+
+    layout.addWidget(sevenTVTokenButton);
+    layout.addWidget(FFZTokenButton);
+    layout.addWidget(BTTVTokenButton);
+
     layout.addTitle("Emotes");
     layout.addCheckbox("Enable Homies global emotes (requires restart)",
                        s.enableHomiesGlobalEmotes);
@@ -66,11 +112,13 @@ void HomiesPage::initLayout(GeneralPageView &layout)
                        s.enableFFZCompletion);
     layout.addCheckbox("Mention users with an at sign (@User)",
                        s.mentionUsersWithAt);
-    layout.addCheckbox("Automatically join separated links (http<s>:/ / → http<s>://)",
-                       s.joinSeparatedLinks);
+    layout.addCheckbox(
+        "Automatically join separated links (http<s>:/ / → http<s>://)",
+        s.joinSeparatedLinks);
 
-    layout.addCheckbox("Automatically separate links (http<s>:// → http<s>:/ /)",
-                       s.separateLinks);
+    layout.addCheckbox(
+        "Automatically separate links (http<s>:// → http<s>:/ /)",
+        s.separateLinks);
 
     QPushButton *selectChannelsSeparateLinksButton = new QPushButton();
     {
@@ -79,13 +127,16 @@ void HomiesPage::initLayout(GeneralPageView &layout)
     }
 
     s.separateLinks.connect(
-            [selectChannelsSeparateLinksButton](const bool &value, auto) {
-                if (value == 1) selectChannelsSeparateLinksButton->setEnabled(true);
-                else selectChannelsSeparateLinksButton->setEnabled(false);
-    });
+        [selectChannelsSeparateLinksButton](const bool &value, auto) {
+            if (value == 1)
+                selectChannelsSeparateLinksButton->setEnabled(true);
+            else
+                selectChannelsSeparateLinksButton->setEnabled(false);
+        });
 
     connect(selectChannelsSeparateLinksButton, &QPushButton::clicked, [=]() {
-        auto selectChannelsSeparateLinksWidget = new SelectChannelSeparateLinksDialog();
+        auto selectChannelsSeparateLinksWidget =
+            new SelectChannelSeparateLinksDialog();
 
         selectChannelsSeparateLinksWidget->show();
         selectChannelsSeparateLinksWidget->raise();