+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## Homies (#6)
+- Major: Added a button to add emotes from 7TV
 - Major: Added an option to separate and join URLs in selected channels
 - Minor: Added `Search in Google` option when text is selected
 - Minor: Added options to disable autocomplete for all global emotes providers